# MachInst backends: sink constants to use-sites to avoid long live-ranges.