@julian-seward1 2020-05-05
Nice, and looks useful.  This is definitely something we should have.  There are a couple of areas of concern tho:

Please add a top level description of perhaps a paragraph, of roughly how the transformation works:

- what the three passes do

- what the running state is (including the relationship between `remove`, `remove_list` and `insert`)

- what the big-O costs are

It wasn't obvious to me without some code reading, that there are three separate passes: one to collect up all constants, one to park them at their use points, and one to delete the sourcing instructions.  As a consequence I spent some minutes trying to figure out whether the effects of the module were dependent on the order in which the blocks are visited.  Which isn't the case, but that took some reading/inferring.

Another area of not exactly concern, but interest, is how this interacts with my constant-phi-removal pass.  I can't guess; we'll have to try it out.

There's also the issue that this might be a bit indiscriminate.  It would be trivial to not inline "expensive" constants (for whatever definition of "expensive") simply by not putting them into `insert`.  I don't think we want to do that right now, but you add a 1-liner comment to that effect, to make it easy for anyone in future who wants to try that?