@bjorn3 2020-05-05
> and one to delete the sourcing instructions

The MachInst backends just skip lowering of side-effect free instructions without any use of their result, right?