@julian-seward1 2020-05-05
@cfallin I see the following (from `joey_big.clif`):
Before:
```
block31:
     jump block22(v236, v3, v3, v3, v3, v3, v3, v3,
                  /* circa 290 more params, most of which are v3,
                     which is "iconst.i32 0" */)
```
After:
```
block31:
     v53727 = iconst.i32 2
     v53728 = iconst.i32 0
     v53729 = iconst.i32 0
     v53730 = iconst.i32 0
     v53731 = iconst.i32 0
     (290 more such lines)
     jump block22(v53727, v53728, v53729, v53730, v53731, v53732, v53733, v53734, ..)
```
The effect is to generate 290+ moves of zero into a new virtual reg (I checked).  We then jump to the phi translation, where they are all copied into block22's argument vregs.  The post-allocation effect is naturally enough a tidal wave of spills in `block31` followed by a similar number of reloads and then re-spills in the phi translation.

Would it be possible to ensure that the same constant is emitted only once per insn?  I think that would remove two out of the three tidal waves here.  Even special-casing zero would probably help; I suspect that the upstream LLVM (that produces the wasm) does SRA on a bunch of structs, and then is obliged to pass around *some* value even for initially uninitialised struct fields, because the wasm semantics require everything to be initialised.  And so it chooses zero.

Since arbitrary de-duping is O(N^2), we could possibly get most of the win just by tracking the last (eg) 10 values emitted for the current insn, and emit a new `iconst` if it's not in those 10.