@julian-seward1 2020-05-05
Let me just clarify: this transformation is clearly good.  I'm just saying, peering at the output and comparing it with the untransformed version, that it could be even more betterer without much effort.