@julian-seward1 2020-05-05
I also tried to assess how this transformation interacts with constant-phi removal.  Here are some numbers.  Bear in mind these are static counts.  What I take from it is that the two transformations partly but not entirely overlap.  And the best sequence is: constant-phi removal first, then const sinking.
```
joey_small.clif
none        out: insns: 1996 total, 44 spills, 111 reloads, 20 nopzs
phi only    out: insns: 1776 total, 6 spills, 77 reloads, 0 nopzs
sink only   out: insns: 1881 total, 0 spills, 0 reloads, 0 nopzs
phi,sink    out: insns: 1733 total, 0 spills, 0 reloads, 0 nopzs
sink,phi    out: insns: 1733 total, 0 spills, 0 reloads, 0 nopzs

joey_medium.clif
none        out: insns: 7061 total, 211 spills, 923 reloads, 302 nopzs
phi only    out: insns: 5390 total, 66 spills, 439 reloads, 49 nopzs
sink only   out: insns: 6911 total, 157 spills, 566 reloads, 214 nopzs
phi,sink    out: insns: 5336 total, 57 spills, 133 reloads, 39 nopzs
sink,phi    out: insns: 5336 total, 57 spills, 133 reloads, 39 nopzs

joey_big.clif
none        out: insns: 126895 total, 13158 spills, 15417 reloads, 72954 nopzs
phi only    out: insns: 111105 total, 10067 spills, 12242 reloads, 67247 nopzs
sink only   out: insns: 124471 total, 12009 spills, 10458 reloads, 77428 nopzs
phi,sink    out: insns: 107553 total, 8364 spills, 8940 reloads, 70251 nopzs
sink,phi    out: insns: 118704 total, 10579 spills, 9029 reloads, 76691 nopzs
```