@cfallin 2020-05-19
Closing this as the problem is solved now in a better way by #1718.